lib/pyproject.toml
[project]
name = "agent-core"
version = "0.1.0"
description = "Shared library for tool-calling agents"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.66.0",
]

[tool.ruff]
line-length = 100
